+++ src/Plugin/LogMiddlewareInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Freshcells\SoapClientBundle\Plugin;
+
+interface LogMiddlewareInterface
+{
+    public function apply($content): string;
+}