+++ tests/Functional/MiddlewareLogTest.php
@@ -3,20 +3,20 @@
 namespace Freshcells\SoapClientBundle\Tests\Functional;
 
 use Freshcells\SoapClientBundle\SoapClient\SoapClient;
-use Freshcells\SoapClientBundle\Tests\AnonymizeAppKernel;
-use Gamez\Psr\Log\TestLogger;
+use Freshcells\SoapClientBundle\Tests\MiddlewareAppKernel;
+use Psr\Log\Test\TestLogger;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
-class AnonymizeLogTest extends WebTestCase
+class MiddlewareLogTest extends WebTestCase
 {
     protected static function getKernelClass()
     {
-        return AnonymizeAppKernel::class;
+        return MiddlewareAppKernel::class;
     }
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->client = static::createClient([
@@ -36,11 +36,11 @@ class AnonymizeLogTest extends WebTestCase
 
         $this->assertEquals('string', $response->DailyDilbertResult);
 
-        $log = $container->get(TestLogger::class)->log;
+        $log = $container->get(TestLogger::class)->records;
 
         $this->assertCount(3, $log);
-        $this->assertTrue(strpos($log[0]->message->value, '<ADate>*****</ADate>') !== false);
-        $this->assertTrue(strpos($log[2]->message->value, '<DailyDilbertResult>*****</DailyDilbertResult>') !== false);
+        $this->assertTrue(strpos($log[0]['message'], '<ADate>*****</ADate>') !== false);
+        $this->assertTrue(strpos($log[2]['message'], '<DailyDilbertResult>st...</DailyDilbertResult>') !== false);
 
         $container->get('profiler')->get('freshcells_soap_client')->collect(new Request(), new Response());
         $this->assertEquals(1, $container->get('profiler')->get('freshcells_soap_client')->getTotal());